feat: move functions in the main package to separate packages
