+++ api/go.work
@@ -3,5 +3,8 @@ go 1.21.0
 use (
 	.
 	./packages/db
+	./packages/handlers
+	./packages/helpers
+	./packages/types
 	./packages/vin
 )
