+++ api/packages/handlers/go.mod
@@ -1,0 +1,6 @@
+module github.com/gc/handlers
+
+replace "github.com/gc/db" => ../db
+replace "github.com/gc/helpers" => ../helpers
+replace "github.com/gc/types" => ../types
+replace "github.com/gc/vin" => ../vin
